+++ src/entry-client.js
@@ -1,34 +1,45 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import {Provider} from 'react-redux'
+import Loadable from 'react-loadable';
 import {BrowserRouter as Router} from 'react-router-dom';
 
 import App from "./client/App";
 import generatorStore from './client/store';
 
 const store = generatorStore(window.REDUX_STATE)
-const jsx = (
-    <Provider store={ store }>
-        <Router>
-            <App/>
-        </Router>
-    </Provider> 
-);
-ReactDOM.hydrate(jsx, document.getElementById("root"));
 
+
+/**
+ * 
+ * @param {Component} Com 
+ * @param {Boolean} isProd 
+ */
+function render(Com, isProd) {
+    const jsx = (
+        <Provider store={ store }>
+            <Router>
+                <Com/>
+            </Router>
+        </Provider> 
+    );
+
+    if (isProd) {
+        Loadable.preloadReady().then(() => {
+            ReactDOM.hydrate(jsx, document.getElementById("root"));
+        });  
+    } else {
+        ReactDOM.hydrate(jsx, document.getElementById("root"));
+    }
+}
+
+render(App, true);
 
 // 热更新
 if (module.hot) {
-  module.hot.accept("./client/App.js", () => {
-    const App = require("./client/App").default;
-    const jsx = (
-      <Provider store={ store }>
-          <Router>
-              <App/>
-          </Router>
-      </Provider> 
-    );
-
-    ReactDOM.hydrate(jsx, document.getElementById("root"));
-  });
+    module.hot.accept("./client/App.js", () => {
+      const App = require("./client/App").default;
+      render(App, false);
+    });
 }
+