+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-# 更新日志
+
+# 1.0.1 (2018-11-10)
+* 添加 react-loadable
+* 增加 路由组件代码分割，动态加载
 
 
+
+## 存在问题
+* 切换dev 到prod 模式时 ，需要手动修改 src/client/App.js中 对routes的引入